Split submission emails into 2 mailers
Due to the fact we want to intercept and do some special operations on the emails sent to the court, we need a way to differentiate between the user receipt email and the court email.

We can't use the same mailer class as the interceptor work on a whitelist handler (mailer class) basis.

Summarising:

* Renamed `EmailSubmissionMailer` to `SubmissionMailer`, being this a superclass to 2 new subclasess: `ReceiptMailer` and `CourtMailer`. Shared code shifted to the superclass.

* Renamed the email preview class accordingly.

* Organised mail templates under the same directory `app/views/mailers/`

* Specs updated.
