+++ app/views/mailers/receipt_mailer/copy_to_user.en.html.erb
