+++ app/views/mailers/receipt_mailer/copy_to_user.en.text.erb
